Plugins and Fixes
Reworked the Ground Smash, added a delay to the regular gunshot. Added a bunch of plugins mostly regarding camera.

Added some space to the test map as the crow would vanish if they disappeared off the map. Extended the map and locked camera to certain deadzone so it felt like it was still the same map.